refactor take 2
