+++ pkg/container/mirror.go
@@ -1,4 +1,4 @@
-package mirror
+package container
 
 import (
 	"context"
