+++ pkg/container/registry.go
@@ -1,4 +1,4 @@
-package mirror
+package container
 
 import "net/url"
